Support encoding a function, without calling it.
generated functions support "send()" and sendAsync()", but sometimes we
need the encoded function, without calling it.
(e.g. if we want to have multiple signatures, or we want to perform the
actual method send later)
For taht purpose, we added "encodeFunctionCall()", which returns the
encoded function as-is.
the returned byte array can be used by
Web3j.ethCall(createEthCallTrasnaction() ) (for view/pure function),
or can be wrapped with RawTransaction.